requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
pydantic==1.10.15
pytest==8.3.2
httpx==0.27.0
sqlmodel==0.0.21
jinja2==3.1.4
Pillow==10.4.0
